adjusted for 64/32bit

+++ tango/text/stream/LineIterator.d
@@ -98,7 +98,7 @@ class LineIterator(T) : StreamIterator!(T)
                 
         ***********************************************************************/
 
-        protected uint scan (void[] data)
+        protected size_t scan (void[] data)
         {
                 auto content = (cast(T*) data.ptr) [0 .. data.length / T.sizeof];
