+++ tango/text/stream/SimpleIterator.d
@@ -84,7 +84,7 @@ class SimpleIterator(T) : StreamIterator!(T)
 
         ***********************************************************************/
 
-        protected uint scan (void[] data)
+        protected size_t scan (void[] data)
         {
                 auto content = (cast(T*) data.ptr) [0 .. data.length / T.sizeof];
